Enable autoDerivation in a way that works on all versions
