--- core/src/test/scala-3/com/softwaremill/diffx/test/DiffVersionSpecificTest.scala
@@ -1,5 +1,0 @@
-package com.softwaremill.diffx.test
-
-trait DiffVersionSpecificTest { outer: DiffTest =>
-  
-}
